steamate/requirements.txt
Django==4.2
djangorestframework==3.14
gunicorn==20.1.0
psycopg
psycopg2-binary==2.9.10
django-cors-headers==3.14
requests==2.31
python-dotenv==1.0.1
pillow==11.1.0
djangorestframework-simplejwt==5.5.0
social-auth-app-django==5.4.3
langchain_core==0.3.41
langchain_community==0.3.19
langchain_openai==0.3.7
langchain-postgres==0.0.11
pandas==2.2.3
cachetools==5.5.2